Catalog reading improvements
Removed restrictions on extension name of the sequence and annotation
files, but added a warning instead if the extension is not an expected
one. ".fasta" added to expected extensions for sequence file.
Removed restriction on relative paths for input files and folders. If
they have an absolute path, that one will be taken, and if they have a
relative path, it will be considered with regards to the repo location.
Added a verification that all 3 important fields are specified in the
catalog (name, sequence, annotation).
Added unit tests.