Small files improvements
Combined all homology matrices and synteny matrices into a single file
for each to greatly reduce the number of small files generated.
"clusters" folder is now deleted at the end of a node processing as all
information it contained is present in either the node.pep file or
locus_mappings.pkl. This is also to grealty reduce the number of small
files kept.
These changes were made in order to improve manipulation speed of
results for the user. (lots of small files are very slow to manipulate)