Muscle runs without temporary files and fix
Fixed the data written to pickle by the RefineCluster step because it
contained also the pickles of the children.
Made Muscle run using stdin and stdout communicating directly with the
the python code through pipes. Multithreading was limited to a single
thread for now because I need to test more if there are no concurrency
problems between multiple muscle stdin and stdout.
Now needs to be tested more to check whether the output is identical for
bigger clusters (check that there is no input  pipe buffer size issue).