Re-enabled multithreading
Successfully tested input over 64k length size (pipe buffer limit that
deadlocks the thread with certain methods of piping stdin).
Compared results of runs with only 1 thread or with 4 threads, output is
equivalent (only the order in which the consensus sequences are written
to the .pep file is different, but there is a lock to prevent mixing
them).